+++ trade/src/main/java/bisq/trade/bisq_easy/protocol/messages/BisqEasyBtcAddressMessageHandler.java
@@ -24,7 +24,7 @@ import bisq.common.validation.BitcoinAddressValidation;
 import bisq.common.validation.LightningInvoiceValidation;
 import bisq.trade.ServiceProvider;
 import bisq.trade.bisq_easy.BisqEasyTrade;
-import bisq.trade.protocol.events.TradeMessageHandler;
+import bisq.trade.protocol.handler.TradeMessageHandler;
 import lombok.extern.slf4j.Slf4j;
 
 import static com.google.common.base.Preconditions.*;
