+++ trade/src/main/java/bisq/trade/bisq_easy/protocol/messages/BisqEasyRejectTradeMessageHandler.java
@@ -21,7 +21,7 @@ import bisq.common.fsm.Event;
 import bisq.contract.Role;
 import bisq.trade.ServiceProvider;
 import bisq.trade.bisq_easy.BisqEasyTrade;
-import bisq.trade.protocol.events.TradeMessageHandler;
+import bisq.trade.protocol.handler.TradeMessageHandler;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
