+++ trade/src/main/java/bisq/trade/bisq_easy/protocol/messages/BisqEasyReportErrorMessageHandler.java
@@ -20,7 +20,7 @@ package bisq.trade.bisq_easy.protocol.messages;
 import bisq.common.fsm.Event;
 import bisq.trade.ServiceProvider;
 import bisq.trade.bisq_easy.BisqEasyTrade;
-import bisq.trade.protocol.events.TradeMessageHandler;
+import bisq.trade.protocol.handler.TradeMessageHandler;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
