+++ trade/src/main/java/bisq/trade/bisq_easy/protocol/messages/BisqEasyTakeOfferResponseHandler.java
@@ -22,7 +22,7 @@ import bisq.contract.ContractService;
 import bisq.contract.ContractSignatureData;
 import bisq.trade.ServiceProvider;
 import bisq.trade.bisq_easy.BisqEasyTrade;
-import bisq.trade.protocol.events.TradeMessageHandler;
+import bisq.trade.protocol.handler.TradeMessageHandler;
 import bisq.trade.protocol.handler.TradeMessageSender;
 import lombok.extern.slf4j.Slf4j;
