+++ trade/src/main/java/bisq/trade/mu_sig/messages/network/handler/seller_as_maker/MuSigCooperativeClosureMessage_G_Handler.java
@@ -25,7 +25,7 @@ import bisq.trade.protobuf.CloseTradeRequest;
 import bisq.trade.protobuf.MusigGrpc;
 import bisq.trade.mu_sig.messages.grpc.CloseTradeResponse;
 import bisq.trade.mu_sig.messages.network.MuSigCooperativeClosureMessage_G;
-import bisq.trade.protocol.events.TradeMessageHandler;
+import bisq.trade.protocol.handler.TradeMessageHandler;
 import com.google.protobuf.ByteString;
 import lombok.extern.slf4j.Slf4j;
