+++ trade/src/main/java/bisq/trade/mu_sig/messages/network/handler/seller_as_maker/MuSigSetupTradeMessage_C_Handler.java
@@ -31,7 +31,7 @@ import bisq.trade.protobuf.DepositTxSignatureRequest;
 import bisq.trade.protobuf.MusigGrpc;
 import bisq.trade.protobuf.PartialSignaturesRequest;
 import bisq.trade.protobuf.ReceiverAddressAndAmount;
-import bisq.trade.protocol.events.TradeMessageHandler;
+import bisq.trade.protocol.handler.TradeMessageHandler;
 import bisq.trade.protocol.handler.TradeMessageSender;
 import com.google.common.collect.ImmutableMap;
 import lombok.extern.slf4j.Slf4j;
