+++ trade/src/main/java/bisq/trade/mu_sig/messages/network/not_used_yet/MuSigTakeOfferRequestHandler.java
@@ -21,7 +21,7 @@ import bisq.common.fsm.Event;
 import bisq.contract.ContractSignatureData;
 import bisq.trade.ServiceProvider;
 import bisq.trade.mu_sig.MuSigTrade;
-import bisq.trade.protocol.events.TradeMessageHandler;
+import bisq.trade.protocol.handler.TradeMessageHandler;
 import bisq.trade.protocol.handler.TradeMessageSender;
 import lombok.extern.slf4j.Slf4j;
